# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Kotori","description":"REST API for Kotori","contact":{"name":"frohikey","email":"frohikey@gmail.com"},"license":{"name":"MIT","url":"https://raw.githubusercontent.com/kotorihq/kotori-server/master/LICENSE"}},"basePath":"/","paths":{"/api/instance":{"get":{"tags":["Instance"],"summary":"Get instance name","description":"Just a simple method for testing purposes. It returns information about instance which you already know :)","operationId":"ApiInstanceGet","consumes":[],"produces":["text/plain","application/json","text/json"],"responses":{"200":{"description":"Instance information","schema":{"$ref":"#/definitions/Instance"},"examples":{"application/json":{"instance":"master-instance"}}},"500":{"description":"Internal server error"}}}},"/api/projects":{"get":{"tags":["Master"],"operationId":"ApiProjectsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/KotoriCore.Domains.SimpleProject"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]},"post":{"tags":["Master"],"operationId":"ApiProjectsPost","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"name","in":"query","required":false,"type":"string"},{"name":"identifier","in":"query","required":false,"type":"string"}],"responses":{"201":{"description":"Success","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]}},"/api/projects/{projectId}":{"get":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/KotoriCore.Domains.SimpleProject"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]},"put":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdPut","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"name","in":"query","required":false,"type":"string"}],"responses":{"201":{"description":"Success","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]},"delete":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdDelete","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]}},"/api/projects/{projectId}/project-keys":{"get":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdProject-keysGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/KotoriCore.Domains.ProjectKey"}}},"404":{"description":"Not Found","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]}},"/api/projects/{projectId}/project-keys/{key}":{"put":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdProject-keysByKeyPut","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"key","in":"path","required":true,"type":"string"},{"name":"isReadonly","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}},"404":{"description":"Not Found","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]},"post":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdProject-keysByKeyPost","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"key","in":"path","required":true,"type":"string"},{"name":"isReadonly","in":"query","required":false,"type":"boolean"}],"responses":{"201":{"description":"Success","schema":{"type":"string"}},"404":{"description":"Not Found","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]},"delete":{"tags":["Master"],"operationId":"ApiProjectsByProjectIdProject-keysByKeyDelete","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"key","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}},"404":{"description":"Not Found","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Internal server error"}},"security":[{"apiKey":["master"]}]}},"/api/projects/{projectId}/document-types":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocumentType"}}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocumentType"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}/count":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdCountGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"filter","in":"query","required":false,"type":"string"},{"name":"drafts","in":"query","required":false,"type":"boolean"},{"name":"future","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/KotoriServer.Tokens.CountResult"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}/documents/{documentId}/{index}/versions/{version}":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdByIndexVersionsByVersionGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"},{"name":"index","in":"path","required":true,"type":"integer","format":"int64"},{"name":"version","in":"path","required":true,"type":"integer","format":"int64"},{"name":"format","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocument"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}/documents/{documentId}/{index}":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdByIndexGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"},{"name":"index","in":"query","required":false,"type":"integer","format":"int64"},{"name":"format","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocument"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}},"put":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdByIndexPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"},{"name":"index","in":"query","required":false,"type":"integer","format":"int64"},{"name":"content","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"201":{"description":"Success","schema":{"type":"string"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}},"post":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdByIndexPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"},{"name":"index","in":"query","required":false,"type":"integer","format":"int64"},{"name":"content","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"201":{"description":"Success","schema":{"type":"string"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}},"delete":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdByIndexDelete","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"},{"name":"index","in":"query","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/KotoriServer.Tokens.CountResult"}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}/documents/{documentId}/versions":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdVersionsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocumentVersion"}}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}/documents/{documentId}/{index}/versions":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdDocumentsByDocumentIdByIndexVersionsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"documentId","in":"path","required":true,"type":"string"},{"name":"index","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocumentVersion"}}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}},"/api/projects/{projectId}/document-types/{documentType}/{documentTypeId}/find":{"get":{"tags":["Project"],"operationId":"ApiProjectsByProjectIdDocument-typesByDocumentTypeByDocumentTypeIdFindGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"},{"name":"documentType","in":"path","required":true,"type":"string"},{"name":"documentTypeId","in":"path","required":true,"type":"string"},{"name":"top","in":"query","required":false,"type":"integer","format":"int32"},{"name":"select","in":"query","required":false,"type":"string"},{"name":"filter","in":"query","required":false,"type":"string"},{"name":"orderBy","in":"query","required":false,"type":"string"},{"name":"drafts","in":"query","required":false,"type":"boolean"},{"name":"future","in":"query","required":false,"type":"boolean"},{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"format","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/KotoriCore.Domains.SimpleDocument"}}},"404":{"description":"Not Found","schema":{"type":"string"}},"500":{"description":"Internal server error"}}}}},"definitions":{"Instance":{"description":"Instance result.","type":"object","properties":{"instance":{"description":"Gets or sets the instance.","type":"string"}}},"KotoriCore.Domains.SimpleProject":{"type":"object","properties":{"name":{"type":"string"},"identifier":{"type":"string"}}},"KotoriCore.Domains.ProjectKey":{"type":"object","properties":{"key":{"type":"string"},"isReadonly":{"type":"boolean"}}},"KotoriCore.Domains.SimpleDocumentType":{"type":"object","properties":{"identifier":{"type":"string"},"type":{"type":"string"}}},"KotoriServer.Tokens.CountResult":{"description":"Count result.","type":"object","properties":{"count":{"format":"int64","description":"Gets or sets the count.","type":"integer"}}},"KotoriCore.Domains.SimpleDocument":{"type":"object","properties":{"identifier":{"type":"string"},"slug":{"type":"string"},"meta":{"type":"object"},"content":{"type":"string"},"date":{"format":"date-time","type":"string"},"modified":{"format":"date-time","type":"string"},"draft":{"type":"boolean"},"version":{"format":"int64","type":"integer"},"filename":{"type":"string"}}},"KotoriCore.Domains.SimpleDocumentVersion":{"type":"object","properties":{"version":{"format":"int64","type":"integer"},"hash":{"type":"string"},"date":{"format":"date-time","type":"string"}}}},"securityDefinitions":{"apiKey":{"name":"apiKey","in":"header","type":"apiKey","description":"API Key Authentication"}}}